+++ gaios/Shared/Controllers/DialogGroupList/DialogGroupListViewController.swift
@@ -1,0 +1,195 @@
+import Foundation
+import UIKit
+
+protocol DialogGroupListViewControllerDelegate: AnyObject {
+    func didSelectIndexPath(_ indexPath: IndexPath, with type: DialogGroupType)
+}
+
+class DialogGroupListViewController: UIViewController {
+
+    @IBOutlet weak var tappableBg: UIView!
+    @IBOutlet weak var handle: UIView!
+    @IBOutlet weak var anchorBottom: NSLayoutConstraint!
+    @IBOutlet weak var lblTitle: UILabel!
+
+    @IBOutlet weak var cardView: UIView!
+    @IBOutlet weak var scrollView: UIScrollView!
+
+    @IBOutlet weak var tableView: UITableView!
+    @IBOutlet weak var tableHeight: NSLayoutConstraint!
+
+    weak var delegate: DialogGroupListViewControllerDelegate?
+
+    var viewModel: DialogGroupListViewModel!
+
+    let headerH = 36.0
+
+    lazy var blurredView: UIView = {
+        let containerView = UIView()
+        let blurEffect = UIBlurEffect(style: .dark)
+        let customBlurEffectView = CustomVisualEffectView(effect: blurEffect, intensity: 0.4)
+        customBlurEffectView.frame = self.view.bounds
+
+        let dimmedView = UIView()
+        dimmedView.backgroundColor = .black.withAlphaComponent(0.3)
+        dimmedView.frame = self.view.bounds
+        containerView.addSubview(customBlurEffectView)
+        containerView.addSubview(dimmedView)
+        return containerView
+    }()
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        setContent()
+        setStyle()
+
+        view.addSubview(blurredView)
+        view.sendSubviewToBack(blurredView)
+
+        ["DialogGroupListCell" ].forEach {
+            tableView.register(UINib(nibName: $0, bundle: nil), forCellReuseIdentifier: $0)
+        }
+
+        view.alpha = 0.0
+        anchorBottom.constant = -cardView.frame.size.height
+
+        let swipeDown = UISwipeGestureRecognizer(target: self, action: #selector(didSwipe))
+            swipeDown.direction = .down
+            self.view.addGestureRecognizer(swipeDown)
+        let tapToClose = UITapGestureRecognizer(target: self, action: #selector(didTap))
+            tappableBg.addGestureRecognizer(tapToClose)
+    }
+
+    @objc func didSwipe(gesture: UIGestureRecognizer) {
+
+        if let swipeGesture = gesture as? UISwipeGestureRecognizer {
+            switch swipeGesture.direction {
+            case .down:
+                dismiss(nil)
+            default:
+                break
+            }
+        }
+    }
+
+    @objc func didTap(gesture: UIGestureRecognizer) {
+
+        dismiss(nil)
+    }
+
+    func setContent() {
+        lblTitle.text = viewModel?.title ?? ""
+    }
+
+    func setStyle() {
+        cardView.layer.cornerRadius = 20
+        cardView.layer.maskedCorners = [.layerMinXMinYCorner, .layerMaxXMinYCorner]
+        handle.cornerRadius = 1.5
+    }
+
+    override func viewDidAppear(_ animated: Bool) {
+        super.viewDidAppear(animated)
+
+        tableHeight.constant = CGFloat( tableView.contentSize.height )
+        anchorBottom.constant = 0
+        UIView.animate(withDuration: 0.3) {
+            self.view.alpha = 1.0
+            self.view.layoutIfNeeded()
+        }
+    }
+
+    func dismiss(_ indexPath: IndexPath?) {
+
+        anchorBottom.constant = -cardView.frame.size.height
+        UIView.animate(withDuration: 0.3, animations: {
+            self.view.alpha = 0.0
+            self.view.layoutIfNeeded()
+        }, completion: { _ in
+            self.dismiss(animated: false, completion: nil)
+            if let indexPath = indexPath {
+                self.delegate?.didSelectIndexPath(indexPath, with: self.viewModel.type)
+            }
+        })
+    }
+}
+
+extension DialogGroupListViewController: UITableViewDelegate, UITableViewDataSource {
+
+    func numberOfSections(in tableView: UITableView) -> Int {
+        return viewModel.dataSource.0.count
+    }
+
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return viewModel.rowsInSection(section)
+    }
+
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let modeType = viewModel.modeTypeAt(indexPath)
+        switch modeType {
+        case .simple:
+            if let cell = tableView.dequeueReusableCell(withIdentifier: DialogGroupListCell.identifier, for: indexPath) as? DialogGroupListCell, let model = viewModel.modelAt(indexPath) {
+                cell.configure(model: model, indexPath: indexPath)
+                cell.selectionStyle = .none
+                return cell
+            }
+        }
+        return UITableViewCell()
+    }
+
+    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
+        return UITableView.automaticDimension
+    }
+
+    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+        dismiss(indexPath)
+    }
+
+    func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
+        return headerH
+    }
+
+    func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
+        return headerView(viewModel.sectionName(section))
+    }
+
+    func tableView(_ tableView: UITableView, heightForFooterInSection section: Int) -> CGFloat {
+        return 0.1
+    }
+
+    func tableView(_ tableView: UITableView, viewForFooterInSection section: Int) -> UIView? {
+        return nil
+    }
+}
+
+extension DialogGroupListViewController {
+
+    func headerView(_ txt: String) -> UIView {
+
+        let section = UIView(frame: CGRect(x: 0, y: 0, width: tableView.frame.width, height: headerH))
+        section.backgroundColor = UIColor.clear
+        let title = UILabel(frame: .zero)
+        title.font = .systemFont(ofSize: 12.0, weight: .medium)
+        title.text = txt
+        title.textColor = UIColor.gW60()
+        title.numberOfLines = 0
+        let line = UIView(frame: CGRect(x: 0, y: 0, width: 0, height: 1.0))
+        line.backgroundColor = UIColor.white.withAlphaComponent(0.1)
+        line.translatesAutoresizingMaskIntoConstraints = false
+        title.translatesAutoresizingMaskIntoConstraints = false
+        section.addSubview(title)
+        section.addSubview(line)
+
+        NSLayoutConstraint.activate([
+            title.centerYAnchor.constraint(equalTo: section.centerYAnchor, constant: 10.0),
+            title.leadingAnchor.constraint(equalTo: section.leadingAnchor, constant: 25)
+        ])
+        NSLayoutConstraint.activate([
+            line.centerYAnchor.constraint(equalTo: section.centerYAnchor, constant: 10.0),
+            line.leadingAnchor.constraint(equalTo: title.trailingAnchor, constant: 10),
+            line.trailingAnchor.constraint(equalTo: section.trailingAnchor, constant: -25),
+            line.heightAnchor.constraint(equalToConstant: 1)
+        ])
+        return section
+    }
+}