+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+ 
+### Changed
+- Updates in wallet preferences dialog
+
+
 ### [4.0.30] - 2024-06-13
 
 ### Added